# swagger/sbAPi.yaml
openapi: 3.0.0
info:
  title: SafeBites OpenAPI 3.0
  description: API for managing and researching restaurants
  license:
    name: MIT
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/filippo.bassetti/SafeBitesApi/1.0.0
  description: SwaggerHub API Auto Mocking
- url: http://localhost:8000/api/v1
  description: Localhost
paths:
...
  /rating/{rest_id}:
    get:
      summary: View all ratings for a restaurant
      description: "Retrieves a list of all ratings for the specified restaurant. Optionally, filter by the user who left the rating."
      parameters:
      - name: rest_id
        in: path
        description: The ID of the rated restaurant.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: user_id
        in: query
        description: The ID of the user who left the rating.
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: A collection of ratings (filtered by restaurant and user if
            provided).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rating'
        "404":
          description: No ratings found for the specified restaurant.
    put:
      summary: Update a rating for a restaurant by a user
      description: Updates the rating for the specified restaurant made by the user.
      parameters:
      - name: rest_id
        in: path
        description: The ID of the rated restaurant.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: user_id
        in: query
        description: The ID of the user who left the rating.
        required: true
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rating'
        required: true
      responses:
        "200":
          description: Rating updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rating'
        "404":
          description: Rating not found.
        "400":
          description: Invalid input.
    delete:
      summary: Delete a rating for a restaurant by a user
      description: Deletes the rating for the specified restaurant made by the user.
      parameters:
      - name: rest_id
        in: path
        description: The ID of the rated restaurant.
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: user_id
        in: query
        description: The ID of the user who left the rating.
        required: true
        style: form
        explode: true
        schema:
          type: string
      responses:
        "204":
          description: Rating deleted successfully (No Content).
        "404":
          description: No matching rating found for the specified restaurant and user.
  /rating:
    post:
      summary: Create a rating for a restaurant
      description: Creates a new rating for a restaurant by a user.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Rating'
        required: true
      responses:
        "201":
          description: Rating created. The Location header contains a link to the
            newly created rating.
          headers:
            Location:
              description: Link to the newly created rating.
              style: simple
              explode: false
              schema:
                type: string
...
components:
  schemas:
...
    Rating:
      required:
      - rating
      - rest_id
      - user_id
      type: object
      properties:
        rest_id:
          type: string
          description: ID of the rated restaurant.
        user_id:
          type: string
          description: ID of the user that left the rating.
        rating:
          maximum: 5
          minimum: 0
          type: integer
          description: Rating given by the user.
...